pyproject.toml
[tool.poetry]
name = "python-depends"
packages = [
    { include="depends", from="." }
]
version = "0.1.0"
description = "A dependency injector as in FastAPI"
repository = "https://github.com/Dimfred/depends.git"
homepage = "https://github.com/Dimfred/depends"
license = "MIT"
authors = ["Dmitrij Vinokour <dimfred.1337@web.de>"]

[tool.poetry.dependencies]
python = "^3.8"
python-acache = {git = "https://github.com/Dimfred/acache.git", rev = "main"}

[tool.poetry.dev-dependencies]
pytest = "^7.1.3"
pytest_async = "^0.1.1"
pytest-cov = "^3.0.0"
pytest-asyncio = "^0.19.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
